[fix](ABC-722): Add iconSVGCreator to create icon svg, remove await in activity timeline and change to call one method in iconUtils
