+++ src/modules/base/iconUtils/iconUtils.js
@@ -32,6 +32,7 @@
 
 import { getPathPrefix, getToken } from 'lightning/configProvider';
 import isIframeInEdge from './isIframeInEdge';
+import { IconSVGCreator } from './iconSVGCreator'
 
 const validNameRe = /^([a-zA-Z]+):([a-zA-Z]\w*)$/;
 const underscoreRe = /_/g;
@@ -118,3 +119,34 @@ export const computeSldsClass = (iconName) => {
 };
 
 export { polyfill } from './polyfill';
+
+let iconCreator;
+
+// SALESFORCE ?
+// export const createSVGIcon = (iconInformation, foreignObjectForIcon, resetMethodOfView) => {
+//     if(!iconCreator){
+//         iconCreator = new IconSVGCreator(resetMethodOfView); 
+//     }
+
+//     iconInformation.xLinkHref = getIconPath(`${iconInformation.category}:${iconInformation.iconName}`);
+//     return iconCreator.createIconFromDefaultPath(foreignObjectForIcon, iconInformation);
+// }
+
+export const createSVGIcon = (iconInformation, foreignObjectForIcon, resetMethodOfView) => {
+    if(!iconCreator){
+        iconCreator = new IconSVGCreator(resetMethodOfView); 
+    }
+
+    let iconSVG;
+    if (!iconCreator.isIconPathAvailable) {
+        iconSVG = iconCreator.createIconFromTemplate(foreignObjectForIcon, iconInformation);
+    } else { 
+        iconInformation.xLinkHref = iconCreator.getIconXLinkHref(iconInformation.category, iconInformation.iconName); 
+        iconSVG = iconCreator.createIconFromDefaultPath(
+            foreignObjectForIcon,
+            iconInformation
+        );
+    }
+
+    return iconSVG;
+}
